fix: サイズ計測の処理をカスタムフックに切り出す
